refactor(experiment_data): collect experiment data related structures in new module
This makes the datastructures accessible without having to import the repository and
its imports. At the same time structures are kept together which change together.
