+++ src/icon/server/data_access/experiment_library_client.py
@@ -10,8 +10,8 @@ if TYPE_CHECKING:
     from icon.server.api.models.parameter_metadata import (
         ParameterMetadata,
     )
-    from icon.server.data_access.db_context.influxdb_v1 import DatabaseValueType
-    from icon.server.data_access.repositories.experiment_data_repository import (
+    from icon.server.data_access.experiment_data import (
+        DatabaseValueType,
         ReadoutMetadata,
     )
