+++ src/icon/server/data_access/pycrystal_experiment_library_client.py
@@ -27,12 +27,12 @@ if TYPE_CHECKING:
     from icon.server.api.models.experiment_dict import (
         ExperimentDict,
     )
-    from icon.server.data_access.db_context.influxdb_v1 import DatabaseValueType
-    from icon.server.data_access.experiment_library_client import ParameterMetadataDict
-    from icon.server.data_access.repositories.experiment_data_repository import (
+    from icon.server.data_access.experiment_data import (
+        DatabaseValueType,
         PlotWindowMetadata,
         ReadoutMetadata,
     )
+    from icon.server.data_access.experiment_library_client import ParameterMetadataDict
 
 logger = logging.getLogger("experiment_library")
 logging.getLogger("pycrystal").setLevel(logging.ERROR)
