+++ src/icon/server/data_access/reconfigurable_experiment_library_client.py
@@ -13,11 +13,9 @@ from icon.server.data_access.experiment_library_client import (
 )
 
 if TYPE_CHECKING:
-    from icon.server.api.models.experiment_dict import (
-        ExperimentDict,
-    )
-    from icon.server.data_access.db_context.influxdb_v1 import DatabaseValueType
-    from icon.server.data_access.repositories.experiment_data_repository import (
+    from icon.server.api.models.experiment_dict import ExperimentDict
+    from icon.server.data_access.experiment_data import (
+        DatabaseValueType,
         ReadoutMetadata,
     )
