+++ src/icon/server/data_access/venv_experiment_library_client.py
@@ -8,11 +8,11 @@ from icon.server.data_access.venv_exec import VirtualEnvironment
 
 if TYPE_CHECKING:
     from icon.server.api.models.experiment_dict import ExperimentDict
-    from icon.server.data_access.db_context.influxdb_v1 import DatabaseValueType
-    from icon.server.data_access.experiment_library_client import ParameterMetadataDict
-    from icon.server.data_access.repositories.experiment_data_repository import (
+    from icon.server.data_access.experiment_data import (
+        DatabaseValueType,
         ReadoutMetadata,
     )
+    from icon.server.data_access.experiment_library_client import ParameterMetadataDict
 
 venv_logger = logging.getLogger("venv")
