+++ src/icon/server/fitting/__init__.py
@@ -1,4 +1,5 @@
-from icon.server.fitting.fit_runner import FitResult, run_curve_fit
+from icon.server.data_access.experiment_data import FitResult
+from icon.server.fitting.fit_runner import run_curve_fit
 from icon.server.fitting.models import FIT_MODELS, FitFunctionType, FitModel
 
 __all__ = [
