+++ src/icon/server/hardware_processing/worker.py
@@ -12,14 +12,14 @@ import pytz
 import socketio.exceptions
 
 from icon.config.config import get_config
+from icon.server.data_access.experiment_data import (
+    ExperimentDataPoint,
+)
 from icon.server.data_access.models.enums import DeviceStatus, JobRunStatus
 from icon.server.data_access.models.sqlite.scan_parameter import (
     contains_realtime_parameter,
 )
 from icon.server.data_access.repositories.device_repository import DeviceRepository
-from icon.server.data_access.repositories.experiment_data_repository import (
-    ExperimentDataPoint,
-)
 from icon.server.data_access.repositories.job_run_repository import JobRunRepository
 from icon.server.hardware_processing.utils import extract_hardware_error_message
 from icon.server.post_processing.task import PostProcessingTask
@@ -28,7 +28,7 @@ from icon.server.utils.handle_keyboard_interrupt import handle_keyboard_interrup
 if TYPE_CHECKING:
     import queue
 
-    from icon.server.data_access.db_context.influxdb_v1 import DatabaseValueType
+    from icon.server.data_access.experiment_data import DatabaseValueType
     from icon.server.data_access.models.sqlite.device import Device
     from icon.server.hardware_processing.hardware_controller import HardwareController
     from icon.server.hardware_processing.task import HardwareProcessingTask
@@ -205,14 +205,14 @@ class HardwareProcessingWorker(multiprocessing.Process):
                 timestamp = datetime.now(timezone)
                 self._hardware_controller.send(data=task.sequence_json.encode("utf-8"))
                 self._hardware_controller.run()
-                result = self._hardware_controller.receive()
+                readouts = self._hardware_controller.receive()
 
                 experiment_data_point = ExperimentDataPoint(
                     index=task.data_point_index,
                     scan_params=task.scanned_params,
-                    result_channels=result.result_channels,
-                    shot_channels=result.shot_channels,
-                    vector_channels=result.vector_channels,
+                    result_channels=readouts.result_channels,
+                    shot_channels=readouts.shot_channels,
+                    vector_channels=readouts.vector_channels,
                     timestamp=timestamp.isoformat(),
                     sequence_json=task.sequence_json,
                 )
