+++ src/icon/server/hardware_processing/zedboard_controller.py
@@ -10,7 +10,7 @@ except ImportError:
 
 
 from icon.config.config import get_config
-from icon.server.data_access.repositories.experiment_data_repository import ResultDict
+from icon.server.data_access.experiment_data import Readouts
 from icon.server.hardware_processing.hardware_controller import (
     HardwareController,
     StatusFlag,
@@ -61,10 +61,10 @@ class ZedboardController(HardwareController):
     def run(self) -> None:
         self._zedboard.sequence_JSON_parser.Parse_JSON_Header()  # type: ignore
 
-    def receive(self) -> ResultDict:
+    def receive(self) -> Readouts:
         results: tiqi_zedboard.zedboard.Result = self._zedboard.sequence_JSON_parser()  # type: ignore
 
-        return ResultDict(
+        return Readouts(
             result_channels=results.result_channels,
             vector_channels=results.vector_channels
             if results.vector_channels is not None
