+++ src/icon/server/utils/scannable_device_parameters.py
@@ -9,7 +9,7 @@ from pydase.utils.serialization.serializer import (
 )
 from pydase.utils.serialization.types import SerializedObject
 
-from icon.server.data_access.db_context.influxdb_v1 import DatabaseValueType
+from icon.server.data_access.experiment_data import DatabaseValueType
 from icon.server.web_server.socketio_emit_queue import emit_queue
